Add new emailconf.php
emailconf,php is a template to send an email (text or html) confirmation
to an email address passed as parameter. The data to be entered is the
From address, the smtp server, smpt username and password, port. You can
get this information from your hosting provider.
You can choose to send an html or text email, and attach a file . If you
attach a file make sure to look up the mime type of the file.